Add new option --no-assume-texthtml
set_content_type assumed that the MIME type of a downloaded file is
text/html if the Content-Type is not provided.

But this is not always desirable, for example to avoid descending into
non-HTML files when recursively downloading.

Provide an option --no-assume-texthtml to disable this assumption, and
wgetrc option assume-texthtml where this can be set explicitly.

 * src/http.c (set_content_type): conditionally assume text/html
 * doc/wget.texi: add --no-assume-texthtml and assume-texthtml config
 * src/init.c: initialize assume_texthtml option
 * src/main.c: add assume-texthtml option
 * src/options.h: add assume_texthtml option

Note that this is not an issue in wget2, which assumes MIME type
application/octet-stream when Content-Type is not provided.